+++ src/datatables-filterdropdown.js
@@ -1,7 +1,7 @@
 /*!
  * DataTables - filterDropDown plugin (modernized fork by Peter Pfeufer)
  *
- * @version 0.0.5
+ * @version 0.0.6
  * @author Peter Pfeufer
  * @license GPL-3.0 or later
  * @link https://github.com/ppfeufer/datatables-filterdropdown
@@ -69,7 +69,7 @@
 })((window, document, DataTable) => { // jshint ignore:line
     'use strict';
 
-    const version = '0.0.5';
+    const version = '0.0.6';
 
     /**
      * Default settings for the filterDropDown plugin.