+++ CHANGELOG.md
@@ -48,6 +48,8 @@ Section Order:
 
 <!-- Your changes go here -->
 
+## [0.0.6] - 2026-09-24
+
 ### Fixed
 
 - `maxWidth` option was not working as expected, now it is properly limiting the
@@ -105,6 +107,7 @@ Section Order:
 [0.0.3]: https://github.com/ppfeufer/datatables-filterdropdown/compare/v0.0.2...v0.0.3 "v0.0.3"
 [0.0.4]: https://github.com/ppfeufer/datatables-filterdropdown/compare/v0.0.3...v0.0.4 "v0.0.4"
 [0.0.5]: https://github.com/ppfeufer/datatables-filterdropdown/compare/v0.0.4...v0.0.5 "v0.0.5"
-[in development]: https://github.com/ppfeufer/datatables-filterdropdown/compare/v0.0.5...HEAD "In Development"
+[0.0.6]: https://github.com/ppfeufer/datatables-filterdropdown/compare/v0.0.5...v0.0.6 "v0.0.6"
+[in development]: https://github.com/ppfeufer/datatables-filterdropdown/compare/v0.0.6...HEAD "In Development"
 [keep a changelog]: http://keepachangelog.com/ "Keep a Changelog"
 [semantic versioning]: http://semver.org/ "Semantic Versioning"